Atualizações no master detail, não permitir repetir valores de uma mesma coluna e extração de dados enviados

+++ Lib/Pixel/Form/MasterDetail/MasterDetail.php
@@ -40,6 +40,15 @@ use Pixel\Arquivo\ArquivoUpload;
 class MasterDetail
 {
 
+    private $dados;
+    private $contaRepeticao;
+
+    public function __construct()
+    {
+        $this->dados = [];
+        $this->contaRepeticao = [];
+    }
+
     public function gravar(FormMasterDetail $config)
     {
         $identifica = $config->getIdentifica();
@@ -61,12 +70,19 @@ class MasterDetail
         $itens = \filter_input(\INPUT_POST, 'sisMasterDetailIten' . $nome, \FILTER_DEFAULT, \FILTER_REQUIRE_ARRAY);
         $confHidden = \json_decode(\str_replace('\'', '"', \filter_input(\INPUT_POST, 'sisMasterDetailConf' . $nome, \FILTER_DEFAULT)));
 
-        $doBanco = \explode(',', $confHidden->ativos);
+        if ($confHidden) {
+            $doBanco = \explode(',', $confHidden->ativos);
+        } else {
+            $doBanco = [];
+        }
         $ativos = [];
 
 
         $coringas = [];
         $coringasMaster = [];
+
+        $this->dados = [];
+
         foreach ($itens as $coringa) {
 
             if ($coringa == $confHidden->coringa) {
@@ -85,6 +101,10 @@ class MasterDetail
 
             $coringas[] = $coringa;
         }
+
+        print_r($this->contaRepeticao);
+
+        $config->setDados($this->dados);
 
         $objPai = $config->getObjetoPai();
 
@@ -125,10 +145,29 @@ class MasterDetail
 
         $colunasCrud = [];
         $grupo = [];
+        $naoRepetir = $config->getNaoRepetir();
+
         foreach ($campos as $campo => $objForm) {
 
             $objForm->setNome($campo);
-            $objForm->setValor($objPai->retornaValor($campo . $coringa));
+            $valorCampo = $objPai->retornaValor($campo . $coringa);
+            $objForm->setValor($valorCampo);
+
+            if ($naoRepetir and \in_array($campo, $naoRepetir)) {
+
+                if (\array_key_exists($campo, $this->contaRepeticao)) {
+
+                    if (\in_array($valorCampo, $this->contaRepeticao[$campo])) {
+                        throw new \Exception($objForm->getIdentifica() . ' - não pode ser repetido!');
+                    }
+
+                    $this->contaRepeticao[$campo][] = $valorCampo;
+                } else {
+                    $this->contaRepeticao[$campo][] = $valorCampo;
+                }
+            }
+
+            $this->dados[$coringa][$campo] = $valorCampo;
 
             if ($objForm->getTipoBase() === 'upload') {
                 $objForm->setNome($objForm->getNome() . $coringa);
@@ -143,7 +182,9 @@ class MasterDetail
 
         $objPai->validar();
 
-        $crudUtil->update($tabela, $colunasCrud, $objPai, [$codigo => $coringa], [], ['upload']);
+        if ($config->getGravar()) {
+            $crudUtil->update($tabela, $colunasCrud, $objPai, [$codigo => $coringa], [], ['upload']);
+        }
     }
 
     private function insert($config, $coringa)
@@ -153,16 +194,35 @@ class MasterDetail
         $tabela = $config->getTabela();
         $campoReferencia = $config->getCampoReferencia();
         $codigoReferencia = $config->getCodigoReferencia();
-        
+
         $campos = $config->getCampos();
         $objPai = $config->getObjetoPai();
 
         $colunasCrud = [];
         $grupo = [];
+        $naoRepetir = $config->getNaoRepetir();
+
         foreach ($campos as $campo => $objForm) {
 
             $objForm->setNome($campo);
-            $objForm->setValor($objPai->retornaValor($campo . $coringa));
+            $valorCampo = $objPai->retornaValor($campo . $coringa);
+            $objForm->setValor($valorCampo);
+
+            if ($naoRepetir and \in_array($campo, $naoRepetir)) {
+
+                if (\array_key_exists($campo, $this->contaRepeticao)) {
+
+                    if (\in_array($valorCampo, $this->contaRepeticao[$campo])) {
+                        throw new \Exception($objForm->getIdentifica() . ' - não pode ser repetido!');
+                    }
+
+                    $this->contaRepeticao[$campo][] = $valorCampo;
+                } else {
+                    $this->contaRepeticao[$campo][] = $valorCampo;
+                }
+            }
+
+            $this->dados[$coringa][$campo] = $valorCampo;
 
             if ($objForm->getTipoBase() === 'upload') {
                 $objForm->setNome($objForm->getNome() . $coringa);
@@ -176,20 +236,23 @@ class MasterDetail
         $objPai->processarForm($grupo);
 
         $objPai->validar();
-        
+
         //Crud Extra
         $crudExtra = $config->getCrudExtra();
-        
-        if($crudExtra){
-            foreach ($crudExtra as $confExtra){
+
+        if ($crudExtra) {
+            foreach ($crudExtra as $confExtra) {
                 $colunasCrud[] = $confExtra[0];
                 $objPai->set($confExtra[0], $confExtra[1], $confExtra[2]);
             }
-        } 
-        
+        }
+
         $colunasCrud[] = $campoReferencia;
         $objPai->set($campoReferencia, $codigoReferencia, 'numero');
-        return $crudUtil->insert($tabela, $colunasCrud, $objPai, ['upload']);
+
+        if ($config->getGravar()) {
+            return $crudUtil->insert($tabela, $colunasCrud, $objPai, ['upload']);
+        }
     }
 
     private function removeItens(FormMasterDetail $config, array $aRemover = [])
@@ -220,7 +283,9 @@ class MasterDetail
                     $objetoRemover->{$metodoRemover}($dados[$codigo]);
                 }
 
-                $crudUtil->delete($tabela, [$codigo => $dados[$codigo]]);
+                if ($config->getGravar()) {
+                    $crudUtil->delete($tabela, [$codigo => $dados[$codigo]]);
+                }
             }
         }
     }
